# Core: Faster Token.stringify